requirements.txt
twisted
PyOpenSSL
service_identity
autobahn
dissononce
consonance
python-axolotl
protobuf
python-magic
pillow
setuptools-rust
